Generate Inc/Dec Counter as 16-bit width.

+++ trafgen_conf.h
@@ -18,7 +18,7 @@ enum csum {
 
 struct counter {
 	int type;
-	uint8_t min, max, inc, val;
+	uint16_t min, max, inc, val;
 	off_t off;
 };
